Refactor of sources files to put inside the folder parsers
