+++ finances/importer/parsers/__init__.py
@@ -1,0 +1,4 @@
+from .excel_file import ExcelSourceFile
+from .html_file import HtmlSourceFile
+from .csv_file import CsvSourceFile
+from .qif_file import QifFile
